wifi: fix issue preventing module from joining WPA mixed-mode (TKIP/CCMP) networks
On a mixed mode network, it's possible to use CCMP (pairwise cipher) for all encryption, CCMP for decryption of unicast frames, and TKIP (group cipher) for decryption of broadcast frames.  Due to a bug, the stack wasn't able to find the pairwise cipher for sending its initial frames (typically DHCP or ARP discovery) and would use the group cipher instead.  On an all-CCMP (WPA2) or all-TKIP (WPA) network, the pairwise and group ciphers are the same, and were therefore unaffected by the bug.

Resolution is to make use of the pending pairwise cipher in __wifig_sendpacket(), but only if we aren't sending an EAPOL handshake frame.

Resolves issue RCM-109.