Handle netport interrupt so transmit errors don't hold buffers forever
Excessive collisions end up dropping frames. We don't know that unless we check. So, handle the netport interrupt, and take a hard look at the error possibilities to figure out if we need to halt a dma transaction and release a buffer.  Otherwise, the buffer would never get released and we'd be stuck thinking we still are transmitting a frame.