ppp: [DC-255] add compile-time macro IPCP_DISABLE_IPADDR_REUSE to control IP address negotiation
With macro defined, PPP stack won't use the IP address of another interface in its initial request to the peer.  In some instances of ETH+PPP, the Rabbit would request the IP address used on the Ethernet interface and the peer would drop the connection instead of requesting another address.