aes: maintain Px/Px' register ordering on Rabbit 6000
Because some of the new encryption instructions for the Rabbit 6000
require strict ordering of the pointer and alternate pointer registers,
it is important to retain the "from reset" arrangement of the
standard/alternate registers.  The setEXPCanonical() function ensures
this, but when using task switching this ordering can be disturbed if
two or more tasks use these instructions and a lower priority task is
preempted during encryption operations.  This commit adds guards to all
uses of the EXP instruction to ensure they are used in pairs to retain
the ordering.  In addition, interrupt prologues and epilogues are
adjusted for the Rabbit 6000 to ensure an even number of swaps.

Some code sequences have been altered to avoid the use of EXP when
ALTD/ALTS options are available.

Note that the uC/OS-II context switching code needs to be altered as
well but this does not reside in this repository. See ucos2-patch2.patch
for the necessary changes to Lib/Rabbit4000/UCOS2/UCOS2.LIB.