samples/extract_flash: update for RCM4200, RCM4120
v10.1.0 of the sample that downloads a copy of the firmware stored on
a Rabbit core module.  Updated for the memory map used on boards that
copy firmware from flash to a "fast SRAM" for execution.

All of those boards use 512KB flash, with 512KB memory banks, so there's
no need to map the flash to quadrant 3.  Just dump the full 512KB from
quadrant 4, starting at 3*512KB.